从26.1.2移植

+++ src/main/java/com/ytgld/moonstone/item/ms/maulice/Brain.java
@@ -23,6 +23,7 @@ public class Brain extends MLS {
     public Brain(Properties properties) {
         super(properties);
     }
+
     @ConfigPlugin
     public static class ConfigItem implements RegisterItemConfig {
         public static ModConfigSpec.DoubleValue intValue;
@@ -51,12 +52,12 @@ public class Brain extends MLS {
     }
 
     public static void brainLHurt(LivingDamageEvent.Pre event) {
-        if (event.getSource().getEntity() instanceof Player player){
-            if (Handler.hascurio(player, Items.brain.get())){
+        if (event.getSource().getEntity() instanceof Player player) {
+            if (Handler.hascurio(player, Items.brain.get())) {
                 String name = event.getEntity().getName().getString();
-                player.getPersistentData().putInt(name, player.getPersistentData().getIntOr(name,0) +1);
-                if (player.getPersistentData().getIntOr(name,0)>= ConfigItem.intValue.get().intValue()){
-                    event.setNewDamage((float) (event.getNewDamage() * ConfigItem.intValue2.get().floatValue()));
+                player.getPersistentData().putInt(name, player.getPersistentData().getIntOr(name, 0) + 1);
+                if (player.getPersistentData().getIntOr(name, 0) >= ConfigItem.intValue.get().intValue()) {
+                    event.setNewDamage(event.getNewDamage() * ConfigItem.intValue2.get().floatValue());
                     player.level().playSound(null, player.getX(), player.getY(), player.getZ(), SoundEvents.WARDEN_HEARTBEAT, SoundSource.NEUTRAL, 4.5F, 4.1F);
                     player.getPersistentData().remove(name);
                 }
@@ -64,14 +65,17 @@ public class Brain extends MLS {
         }
 
     }
+
     @Override
     public void onEquip(SlotContext slotContext, ItemStack prevStack, ItemStack stack) {
         Handler.stackCreateTag(stack);
     }
+
     @Override
     public void appendHoverText(ItemStack stack, @javax.annotation.Nullable Level level, List<Component> tooltip, TooltipFlag flags) {
         super.appendHoverText(stack, level, tooltip, flags);
         tooltip.add(Component.translatable("item.brain.tool.string").withStyle(ChatFormatting.DARK_GREEN));
     }
-    public static final String brain ="brain_brain";
+
+    public static final String brain = "brain_brain";
 }
